fix: styling work
