+++ src/pages/programDisaggregations/form/CategoriesSelector.module.css
@@ -1,0 +1,8 @@
+.selectorLabel {
+
+}
+
+.selectorContainer {
+    width: 400px;
+    margin-block: 8px;
+}
