+++ src/pages/programDisaggregations/form/ProgramIndicatorMappingSection.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const ProgramIndicatorMappingSection = () => {
+    return <p>placeholder program indicator mapping section</p>
+}
